Move registry.js
